chore clippy

+++ src/center.rs
@@ -104,10 +104,8 @@ impl CenteredFiberData {
     }
 
     pub fn leading_columns(&self) -> String {
-        let mut c_query_start =
-            self.fiber.record.reference_start() as i64 - self.center_position.position;
-        let mut c_query_end =
-            self.fiber.record.reference_end() as i64 - self.center_position.position;
+        let mut c_query_start = self.fiber.record.reference_start() - self.center_position.position;
+        let mut c_query_end = self.fiber.record.reference_end() - self.center_position.position;
         if self.center_position.strand == '-' {
             c_query_start = -c_query_start;
             c_query_end = -c_query_end;
